fix missing coremod
