--- src/main/java/net/tclproject/mysteriumlib/asm/fixes/UselessClass.java
@@ -1,4 +1,0 @@
-package net.tclproject.mysteriumlib.asm.fixes;
-
-public class UselessClass {
-}
