--- src/main/java/xonin/backhand/coremod/transformers/BackhandAccessTransformer.java
@@ -1,12 +1,0 @@
-package xonin.backhand.coremod.transformers;
-
-import java.io.IOException;
-
-import cpw.mods.fml.common.asm.transformers.AccessTransformer;
-
-public class BackhandAccessTransformer extends AccessTransformer {
-
-    public BackhandAccessTransformer() throws IOException {
-        super("theoffhandmod_at.cfg");
-    }
-}
